Updated tests: run "npm run test"
