+++ spec/06_quote_beginning.txt
